[Maps] remove maps_file_upload plugin and fold public folder into file_upload plugin (#90292)
* get geojson working with api/file_upload/import

* remove maps_file_upload server code

* remove common folder

* remove maps_file_upload plugin

* fix tsconfig paths

* rename file_upload plugin in maps tsconfig

* fix file path

* node scripts/build_plugin_list_docs

* fix webpack compile errors

* telemetry schema cleanup, i18n cleanup, limits cleanup

* remove mapsFileUpload from limits.yml

* remove index pattern link test case

* update telemetry/v2/clusters/_stats for new file_upload path
