+++ docs/developer/plugin-list.asciidoc
@@ -439,10 +439,6 @@ using the CURL scripts in the scripts folder.
 |Visualize geo data from Elasticsearch or 3rd party geo-services.
 
 
-|{kib-repo}blob/{branch}/x-pack/plugins/maps_file_upload/README.md[mapsFileUpload]
-|Deprecated - plugin targeted for removal and will get merged into file_upload plugin
-
-
 |{kib-repo}blob/{branch}/x-pack/plugins/maps_legacy_licensing/README.md[mapsLegacyLicensing]
 |This plugin provides access to the detailed tile map services from Elastic.
