+++ packages/kbn-optimizer/limits.yml
@@ -104,4 +104,4 @@ pageLoadAssetSize:
   presentationUtil: 28545
   spacesOss: 18817
   osquery: 107090
-  mapsFileUpload: 23775
+  fileUpload: 25664
