+++ x-pack/plugins/file_upload/kibana.json
@@ -3,6 +3,6 @@
   "version": "8.0.0",
   "kibanaVersion": "kibana",
   "server": true,
-  "ui": false,
-  "requiredPlugins": ["usageCollection"]
+  "ui": true,
+  "requiredPlugins": ["data", "usageCollection"]
 }
