+++ x-pack/plugins/file_upload/public/components/index_settings.js
