+++ x-pack/plugins/file_upload/public/components/json_import_progress.js
@@ -118,9 +118,7 @@ export class JsonImportProgress extends Component {
                 <a
                   data-test-subj="indexManagementNewIndexLink"
                   target="_blank"
-                  href={`${basePath}/app/kibana#/
-                      management/elasticsearch/index_management/indices/
-                      filter/${indexName}`.replace(/\s/g, '')}
+                  href={`${basePath}/app/management/kibana/indexPatterns`}
                 >
                   {i18n.translate('xpack.fileUpload.jsonImport.indexMgmtLink', {
                     defaultMessage: 'Index Management',
