+++ x-pack/plugins/file_upload/public/components/json_upload_and_parse.js
