+++ x-pack/plugins/file_upload/public/get_file_upload_component.ts
