+++ x-pack/plugins/file_upload/public/index.ts
@@ -7,6 +7,11 @@
 
 import { FileUploadPlugin } from './plugin';
 
-export * from './plugin';
+export function plugin() {
+  return new FileUploadPlugin();
+}
 
-export const plugin = () => new FileUploadPlugin();
+export * from '../common';
+
+export { StartContract } from './plugin';
+export { FileUploadComponentProps } from './get_file_upload_component';
