+++ x-pack/plugins/file_upload/public/kibana_services.js
