+++ x-pack/plugins/file_upload/public/plugin.ts
