+++ x-pack/plugins/file_upload/public/util/file_parser.js
