+++ x-pack/plugins/file_upload/public/util/geo_json_clean_and_validate.js
