+++ x-pack/plugins/file_upload/public/util/http_service.js
