+++ x-pack/plugins/file_upload/public/util/indexing_service.test.js
