+++ x-pack/plugins/file_upload/public/util/size_limited_chunking.js
@@ -5,7 +5,7 @@
  * 2.0.
  */
 
-import { MAX_BYTES } from '../../common/constants/file_import';
+const MAX_BYTES = 31457280;
 
 // MAX_BYTES is a good guideline for splitting up posts, but this logic
 // occasionally sizes chunks so closely to the limit, that the remaining content
