+++ x-pack/plugins/maps/kibana.json
@@ -11,7 +11,7 @@
     "features",
     "inspector",
     "data",
-    "mapsFileUpload",
+    "fileUpload",
     "uiActions",
     "navigation",
     "visualizations",
