+++ x-pack/plugins/maps/public/kibana_services.ts
@@ -27,7 +27,7 @@ export const getIndexPatternService = () => pluginsStart.data.indexPatterns;
 export const getAutocompleteService = () => pluginsStart.data.autocomplete;
 export const getInspector = () => pluginsStart.inspector;
 export const getFileUploadComponent = async () => {
-  return await pluginsStart.mapsFileUpload.getFileUploadComponent();
+  return await pluginsStart.fileUpload.getFileUploadComponent();
 };
 export const getUiSettings = () => coreStart.uiSettings;
 export const getIsDarkMode = () => getUiSettings().get('theme:darkMode', false);
