+++ x-pack/plugins/maps/tsconfig.json
@@ -19,7 +19,7 @@
     { "path": "../../../src/plugins/maps_legacy/tsconfig.json" },
     { "path": "../features/tsconfig.json" },
     { "path": "../licensing/tsconfig.json" },
-    { "path": "../maps_file_upload/tsconfig.json" },
+    { "path": "../file_upload/tsconfig.json" },
     { "path": "../saved_objects_tagging/tsconfig.json" },
   ]
 }
