--- x-pack/plugins/maps_file_upload/kibana.json
@@ -1,8 +1,0 @@
-{
-  "id": "mapsFileUpload",
-  "version": "8.0.0",
-  "kibanaVersion": "kibana",
-  "server": true,
-  "ui": true,
-  "requiredPlugins": ["data", "usageCollection"]
-}
