--- x-pack/plugins/maps_file_upload/mappings.ts
@@ -1,16 +1,0 @@
-/*
- * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
- * or more contributor license agreements. Licensed under the Elastic License
- * 2.0; you may not use this file except in compliance with the Elastic License
- * 2.0.
- */
-
-export const mappings = {
-  'file-upload-telemetry': {
-    properties: {
-      filesUploadedTotalCount: {
-        type: 'long',
-      },
-    },
-  },
-};
