--- x-pack/plugins/maps_file_upload/public/index.ts
@@ -1,15 +1,0 @@
-/*
- * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
- * or more contributor license agreements. Licensed under the Elastic License
- * 2.0; you may not use this file except in compliance with the Elastic License
- * 2.0.
- */
-
-import { FileUploadPlugin } from './plugin';
-
-export function plugin() {
-  return new FileUploadPlugin();
-}
-
-export { StartContract } from './plugin';
-export { FileUploadComponentProps } from './get_file_upload_component';
