--- x-pack/plugins/maps_file_upload/server/client/errors.js
@@ -1,12 +1,0 @@
-/*
- * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
- * or more contributor license agreements. Licensed under the Elastic License
- * 2.0; you may not use this file except in compliance with the Elastic License
- * 2.0.
- */
-
-import { boomify } from '@hapi/boom';
-
-export function wrapError(error) {
-  return boomify(error, { statusCode: error.status });
-}
