--- x-pack/plugins/maps_file_upload/server/plugin.js
@@ -1,27 +1,0 @@
-/*
- * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
- * or more contributor license agreements. Licensed under the Elastic License
- * 2.0; you may not use this file except in compliance with the Elastic License
- * 2.0.
- */
-
-import { initRoutes } from './routes/file_upload';
-import { setInternalRepository } from './kibana_server_services';
-import { registerFileUploadUsageCollector, fileUploadTelemetryMappingsType } from './telemetry';
-
-export class FileUploadPlugin {
-  constructor() {
-    this.router = null;
-  }
-
-  setup(core, plugins) {
-    core.savedObjects.registerType(fileUploadTelemetryMappingsType);
-    this.router = core.http.createRouter();
-    registerFileUploadUsageCollector(plugins.usageCollection);
-  }
-
-  start(core) {
-    initRoutes(this.router, core.savedObjects.getSavedObjectsRepository);
-    setInternalRepository(core.savedObjects.createInternalRepository);
-  }
-}
