--- x-pack/plugins/maps_file_upload/server/telemetry/index.ts
@@ -1,9 +1,0 @@
-/*
- * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
- * or more contributor license agreements. Licensed under the Elastic License
- * 2.0; you may not use this file except in compliance with the Elastic License
- * 2.0.
- */
-
-export { registerFileUploadUsageCollector } from './file_upload_usage_collector';
-export { fileUploadTelemetryMappingsType } from './mappings';
