--- x-pack/plugins/maps_file_upload/server/telemetry/mappings.ts
@@ -1,22 +1,0 @@
-/*
- * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
- * or more contributor license agreements. Licensed under the Elastic License
- * 2.0; you may not use this file except in compliance with the Elastic License
- * 2.0.
- */
-
-import { SavedObjectsType } from 'src/core/server';
-import { TELEMETRY_DOC_ID } from './telemetry';
-
-export const fileUploadTelemetryMappingsType: SavedObjectsType = {
-  name: TELEMETRY_DOC_ID,
-  hidden: false,
-  namespaceType: 'agnostic',
-  mappings: {
-    properties: {
-      filesUploadedTotalCount: {
-        type: 'long',
-      },
-    },
-  },
-};
