--- x-pack/plugins/maps_file_upload/tsconfig.json
@@ -1,15 +1,0 @@
-{
-  "extends": "../../../tsconfig.base.json",
-  "compilerOptions": {
-    "composite": true,
-    "outDir": "./target/types",
-    "emitDeclarationOnly": true,
-    "declaration": true,
-    "declarationMap": true
-  },
-  "include": ["common/**/*", "public/**/*", "server/**/*", "mappings.ts"],
-  "references": [
-    { "path": "../../../src/plugins/data/tsconfig.json" },
-    { "path": "../../../src/plugins/usage_collection/tsconfig.json" }
-  ]
-}
