+++ x-pack/plugins/telemetry_collection_xpack/schema/xpack_plugins.json
@@ -2219,13 +2219,6 @@
         }
       }
     },
-    "fileUploadTelemetry": {
-      "properties": {
-        "filesUploadedTotalCount": {
-          "type": "long"
-        }
-      }
-    },
     "maps": {
       "properties": {
         "settings": {
