+++ x-pack/tsconfig.json
@@ -27,7 +27,6 @@
     "plugins/licensing/**/*",
     "plugins/lens/**/*",
     "plugins/maps/**/*",
-    "plugins/maps_file_upload/**/*",
     "plugins/maps_legacy_licensing/**/*",
     "plugins/ml/**/*",
     "plugins/observability/**/*",
@@ -130,7 +129,6 @@
     { "path": "./plugins/lens/tsconfig.json" },
     { "path": "./plugins/license_management/tsconfig.json" },
     { "path": "./plugins/licensing/tsconfig.json" },
-    { "path": "./plugins/maps_file_upload/tsconfig.json" },
     { "path": "./plugins/maps_legacy_licensing/tsconfig.json" },
     { "path": "./plugins/maps/tsconfig.json" },
     { "path": "./plugins/ml/tsconfig.json" },
